Use `long`s instead of `int`s

+++ src/Stripe.net/Entities/Accounts/PayoutSchedule.cs
@@ -5,13 +5,13 @@ namespace Stripe
     public class PayoutSchedule : StripeEntity
     {
         [JsonProperty("delay_days")]
-        public int DelayDays { get; set; }
+        public long DelayDays { get; set; }
 
         [JsonProperty("interval")]
         public string Interval { get; set; }
 
         [JsonProperty("monthly_anchor")]
-        public int MonthlyAnchor { get; set; }
+        public long MonthlyAnchor { get; set; }
 
         [JsonProperty("weekly_anchor")]
         public string WeeklyAnchor { get; set; }
