+++ src/Stripe.net/Entities/Cards/Card.cs
@@ -96,10 +96,10 @@ namespace Stripe
         public string DynamicLast4 { get; set; }
 
         [JsonProperty("exp_month")]
-        public int ExpMonth { get; set; }
+        public long ExpMonth { get; set; }
 
         [JsonProperty("exp_year")]
-        public int ExpYear { get; set; }
+        public long ExpYear { get; set; }
 
         [JsonProperty("fingerprint")]
         public string Fingerprint { get; set; }
