+++ src/Stripe.net/Entities/Sources/SourceCodeVerification.cs
@@ -8,7 +8,7 @@ namespace Stripe
         /// The number of attempts remaining to authenticate the source object with a verification code.
         /// </summary>
         [JsonProperty("attempts_remaining")]
-        public int AttemptsRemaining { get; set; }
+        public long AttemptsRemaining { get; set; }
 
         /// <summary>
         /// The status of the code verification, either pending, succeeded or failed.
